[EdgeTpu] Move parse function to ConfigSetting
- move parse function to ConfigSetting.ts
- change ConfigObject's member configSetting to configType

ONE-vscode-DCO-1.0-Signed-off-by: Seongwon Im <imsw0529@gmail.com>
