+++ src/OneExplorer/ConfigSetting.ts
@@ -14,20 +14,87 @@
  * limitations under the License.
  */
 
-import { LocatorRunner } from "./ArtifactLocator";
+import * as path from "path";
+
+import { Logger } from "../Utils/Logger";
+import { Artifact, LocatorRunner } from "./ArtifactLocator";
 
 export abstract class ConfigSetting {
   baseModelsLocatorRunner: LocatorRunner;
   productsLocatorRunner: LocatorRunner;
   // TODO: make sections for updateBaseModelField method
 
+  /**
+   * @brief Parse base models written in the ini object and return the absolute path.
+   *
+   * @param uri cfg uri is required to calculate absolute path
+   *
+   * ABOUT MULTIPLE BASE MODELS
+   *
+   * onecc doesn't support multiple base models.
+   * However, OneExplorer will show the config node below multiple base models
+   * to prevent a case that users cannot find their faulty config files on ONE explorer.
+   *
+   * TODO Move to backend
+   */
+  public parseBaseModels = (filePath: string, iniObj: object): Artifact[] => {
+    const dir = path.dirname(filePath);
+
+    let locatorRunner = this.baseModelsLocatorRunner;
+
+    let artifacts: Artifact[] = locatorRunner.run(iniObj, dir);
+
+    if (artifacts.length > 1) {
+      // TODO Notify the error with a better UX
+      // EX. put question mark next to the config icon
+      Logger.debug(
+        "OneExplorer",
+        `There are multiple input models in the configuration(${filePath}).`
+      );
+    }
+    if (artifacts.length === 0) {
+      // TODO Notify the error with a better UX
+      // EX. showing orphan nodes somewhere
+      Logger.debug(
+        "OneExplorer",
+        `There is no input model in the configuration(${filePath}).`
+      );
+    }
+
+    // Return as list of uri
+    return artifacts;
+  };
+
+  /**
+   * @brief Find derived models written in the ini object and return the absolute path.
+   *
+   * @param filePath cfg file path is required to calculate absolute path
+   *
+   * TODO Move to backend
+   */
+  public parseProducts = (filePath: string, iniObj: object): Artifact[] => {
+    const dir = path.dirname(filePath);
+
+    let locatorRunner = this.productsLocatorRunner;
+
+    /**
+     * When you add a new product type, please append the ext type to
+     * OneTreeDataProvider.fileWatcher too, to prevent a bug.
+     *
+     * TODO Provide better structure to remove this extra work
+     */
+
+    let artifacts: Artifact[] = locatorRunner.run(iniObj, dir);
+
+    return artifacts;
+  };
+
   constructor() {
     this.baseModelsLocatorRunner = new LocatorRunner();
     this.productsLocatorRunner = new LocatorRunner();
-    this._init();
   }
 
-  private _init(): void {
+  public init(): void {
     this._initBaseModelsLocatorRunner();
     this._initProductsLocatorRunner();
   }
